Hide filter nodes with the `/Internal` suffix and remove `access.conf`
pulse-server, with the proposed patch
https://gitlab.freedesktop.org/pipewire/pipewire/-/merge_requests/2064,
will be able to hide a node when the media class has the `/Internal`
suffix.

We can leverage that to hide our filter nodes. This fixes the issue with
some applications (e.g. CS2) that were still showing the filter nodes.

This commit also removes the `access.conf` file because it's not needed
anymore. We don't reply on permissions to hide nodes.

Co-authored-by: Julian Bouzas <julian.bouzas@collabora.com>
Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>