WIP wireplumber: Group audio streams by the "reaper" ancestor
As part of
https://gitlab.freedesktop.org/pipewire/wireplumber/-/merge_requests/646,
we are not grouping audio streams using cgroups anymore. Instead,
wireplumber is using a specific common ancestor process cmdline.

At the moment we are using `reaper`, but this is only for testing and it
is expected to change before going in production.

Related to https://gitlab.steamos.cloud/holo-team/tasks/-/issues/1173

Co-authored-by: Julian Bouzas <julian.bouzas@collabora.com>
Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>