Group playback nodes by the "reaper" ancestor
As part of the upstream change
https://gitlab.freedesktop.org/pipewire/wireplumber/-/merge_requests/646,
this allows us to group all streams with the same specific ancestor.

We use "reaper" because it is a parent that we always expect to be
present, no matter what type of application we are running.

Part of https://gitlab.steamos.cloud/holo-team/tasks/-/issues/1173

Co-authored-by: Julian Bouzas <julian.bouzas@collabora.com>
Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>