pipewire-confs: Increase the device latency to 512/48000
When recording with the microphone we are experiencing audio output
distortion.

A previous attempt at fixing this was to increase the latency in the
game stream nodes. However, that caused issues with Ori and the Will of
the Wisps.

With this commit we still increase the latency to 512/48000, but this
time the latency is applied to the actual device.

Co-authored-by: Julian Bouzas <julian.bouzas@collabora.com>
Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>