pipewire-confs: Switch to the LV2 plugin for rnnoise
We are already using an LV2 plugin for the `valve_deck_microphone`
filter. Switching the rnnoise to LV2 should make it cheaper to process
instead of combining two filters of different types.

Because we can't set custom VAD parameters with the LV2 plugin, we moved
those settings in `noise-suppression-for-voice 1.10-1.1`:
https://gitlab.steamos.cloud/holo/holo/-/merge_requests/1476

Helps https://gitlab.steamos.cloud/holo-team/tasks/-/issues/1739

Co-authored-by: Julian Bouzas <julian.bouzas@collabora.com>
Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>