pipewire-confs: enable memlocking
This tells the pipewire filter-chain to call memlock so that it
doesn't get SIGKILLed when running expensive noise filters in a
capture-and-playback configuration.

It requires the filter-chain user service to be run with a memlock
ulimit (empirically determined) of about 100MiB,
whch in turn requires a suitable systemd config for
user services.

Co-authored-by: Julian Bouzas <julian.bouzas@collabora.com>
Signed-off-by: Vivek Das Mohapatra <vivek@collabora.com>