pipewire-confs: Reintroduce context properties
We previously dropped these in 1cbe32507befda9bac880ffe6c51d4de7779bc3a,
but on further consideration, the built-in default min quantum of 32,
and default quantum of 1024, are likely not what we want.

Let's pick a default which gives us lower latency, and a min that we are
reasonably confident we can support (so if a game requests less, it
won't fail).

The max quantum is set to 8192 instead of being limited as it previously
was, so that users can shave some power usage in media playback
use-cases.