Set HDMI device suspend timeout to 1 hour
For Fremont, this bumps the delay from the default 5 seconds to one
hour. For Jupiter/Galileo, we avoid disabling suspend altogether in
favour of a long timeout.

This allows us to avoid amplifier startup delays, but does allow the
amplifier to shutdown after a long period of inactivity to save a little
power.